Responsive window, and other fixes

+++ Camera.h
@@ -9,9 +9,9 @@
 class Camera {
 
 public:
-    double x = 0;
-    double y = 0;
-    double z = 0;
+    float x = 0;
+    float y = 0;
+    float z = 0;
 };
 
 Camera camera;
